feat: Establish the initial project structure, core services, and foundational UI components for the application.

+++ src-tauri/src/api/modules/downloader.rs
@@ -69,6 +69,6 @@ pub async fn list_module_files(module_id: &str) -> Result<Vec<String>, AppError>
 #[tauri::command]
 #[specta::specta]
 /// Configures download bandwidth limits
-pub fn set_download_settings(enabled: bool, max_speed: u64) {
+pub fn set_download_settings(enabled: bool, max_speed: u32) {
     downloader::DOWNLOADER.set_limit(enabled, max_speed);
 }
